docs: dont add workspaces with no exports (#1355)
This was previously handled by checking if a workspace used thsy but
when that was removed in #1344, the `gui` workspace was trying to be
added to the docs site even though it had no exports. The gui was never
meant to be in the docs site so it's usage of markdown was breaking the
build in a few ways.