fix(run): correctly run pre/post scripts when main command doesnt exist (#1373)
Previously there was a check so that if there was no `command` then
`run` would bail early with an empty command result. This resulted in
command like `vlt build` not actually running packages that had only
`postinstall` or `preinstall` scripts (such as esbuild).

This also refactors a few things:

- moves some `node-gyp` specific functions to the renamed `node-gyp.ts`
file
- simplifies return values of `runImpl()`

Fixes #1372