gui: focused view consolidation (#1375)
## Overview

Previously the 'focused' view and the 'selected' item view were too
similar, and the UX of having to navigate to the header to essentially
'collapse' the potential parent of an item posed an convoluted
experience. This PR consolidates the `focused-view` into the
`selected-item` but removing the `focused-view` component tree entirely,
and just having a 'collapse' button that essentially just collapses the
parent on the left hand pane.

In addition to this, we can now always show the `dependency-sidebar` on
all views within a `selected-item`.

As a result of the above, because we're always showing the
`dependency-sidebar`, external package packages via the search
experience would've always shown an empty state for that (since we do
not have a graph for external packages). Instead we can parse all the
dependencies from the `manifest` of the `selected-item` (for external
packages) and use the utility `record-to-array` to nicely iterate and
show deps for external packages there. Clicking on one of those deps
will navigate to that external package page now. However, because we do
not have a 'resolved' semver value for those deps, we just navigate to
the 'latest' version of that package (for now) As a follow up we can
work on the correct semver navigation for external package deps.

| Before | After |
| --- | --- |
| <img width="1744" height="1310" alt="Screenshot 2025-12-18 at 12 27
43"
src="https://github.com/user-attachments/assets/921710b9-843e-462d-9072-34c442bfac8f"
/> | <img width="1744" height="1310" alt="Screenshot 2025-12-18 at 12 26
20"
src="https://github.com/user-attachments/assets/dc6b93c4-808a-4c3e-aa97-cd13a9c15cf3"
/> |

## Other changes

### Decorators

This PR also updates how the 'diamond' like decorator is rendered on
intersecting grid patterns by removing the need for having nested
elements with a border radius and secondary background color, by
creating a `Cross` component which we can align at intersecting grid
locations to make styling much easier, and having a cleaner component
tree.

### Header updates

This PR also updates some header styling

### Layout updates

This PR updates the layout on `search` pages to also contain a `sidebar`
component to make the UX cleaner when transitioning between certain
routes

### Changelog
- **feat: adds new utility `record-to-array`**
- **feat: creates a `cross` ui decorator component**
- **feat: removes `focused-view` component tree**
- **feat: remvoves test suites for `focused-view` tree**
- **feat: updates styling on the gui**
- **feat: updates tests**
- **chore: lint and format**